+++ CHANGELOG.md
@@ -7,6 +7,8 @@ CHANGELOG
 ### Latest changes
 
 * VFS-7960 Fixed changing space in file browser during files upload
+* VFS-7730 Added reloading files list after qos/metadata/acl change of file with hardlinks
+* VFS-7371 Added showing progress for copied or moved operation
 * VFS-8031 Added select current directory button in items selector (if applicable)
 * VFS-8212 Fixed not working current directory rename
 * VFS-7930 Added "create incremental archive" in archive context menu